Fix modifier key handling on Windows + added recordKeys function
Previously, the plugin didn't release modifier keys properly on Windows after being held down. This commit changes the way keys are invoked by passing modifier keys as boolean values to the platform function. Now, the modifier keys are invoked in the same function as the key that should be invoked. This fixes the issue with modifier key release on Windows.

Note: Currently, this change breaks support for Linux and macOS